Moved tests to DataSourceWithCacheAndSnapshotTest class
