+++ rskj-core/src/test/java/org/ethereum/datasource/DataSourceWithCacheTest.java
@@ -18,10 +18,10 @@ import static org.mockito.Mockito.*;
 
 public class DataSourceWithCacheTest {
 
-    private static final int CACHE_SIZE = 15;
+    protected static final int CACHE_SIZE = 15;
 
-    private HashMapDB baseDataSource;
-    private DataSourceWithCache dataSourceWithCache;
+    protected HashMapDB baseDataSource;
+    protected DataSourceWithCache dataSourceWithCache;
 
     @Before
     public void setupDataSources() {
@@ -242,24 +242,6 @@ public class DataSourceWithCacheTest {
         for (ByteArrayWrapper removedKey : keysToBatchRemove) {
             assertThat(baseDataSource.get(removedKey.getData()), is(nullValue()));
         }
-    }
-
-    @Test
-    public void checkCacheSnapshotLoadTriggered() {
-        CacheSnapshotHandler cacheSnapshotHandler = mock(CacheSnapshotHandler.class);
-        DataSourceWithCacheAndSnapshot.create(baseDataSource, CACHE_SIZE, cacheSnapshotHandler);
-
-        verify(cacheSnapshotHandler, atLeastOnce()).load(anyMap());
-    }
-
-    @Test
-    public void checkCacheSnapshotSaveTriggered() {
-        CacheSnapshotHandler cacheSnapshotHandler = mock(CacheSnapshotHandler.class);
-        DataSourceWithCache dataSourceWithCache = DataSourceWithCacheAndSnapshot.create(baseDataSource, CACHE_SIZE, cacheSnapshotHandler);
-
-        dataSourceWithCache.close();
-
-        verify(cacheSnapshotHandler, atLeastOnce()).save(anyMap());
     }
 
     private Map<ByteArrayWrapper, byte[]> generateRandomValuesToUpdate(int maxValuesToCreate) {
